Feat: started cleaning up consensus mechanism
