+++ packages/consensus/fcob/consensusmechanism_test.go
@@ -22,7 +22,7 @@ func TestOpinionFormer_Scenario2(t *testing.T) {
 	LikedThreshold = 2 * time.Second
 	LocallyFinalizedThreshold = 2 * time.Second
 
-	consensusProvider := NewConsensusProvider()
+	consensusProvider := NewConsensusMechanism()
 
 	testTangle := tangle.New(tangle.Consensus(consensusProvider))
 	defer testTangle.Shutdown()
@@ -133,7 +133,7 @@ func TestOpinionFormer_Scenario2(t *testing.T) {
 	transactionLiked[transactions["6"].ID()] = true
 	transactionLiked[transactions["8"].ID()] = false
 
-	consensusProvider.PayloadOpinionProvider.Vote().Attach(events.NewClosure(func(transactionID string, initialOpinion opinion.Opinion) {
+	consensusProvider.Events.Vote.Attach(events.NewClosure(func(transactionID string, initialOpinion opinion.Opinion) {
 		t.Log("Voting requested for:", transactionID)
 		txID, err := ledgerstate.TransactionIDFromBase58(transactionID)
 		require.NoError(t, err)
@@ -141,13 +141,13 @@ func TestOpinionFormer_Scenario2(t *testing.T) {
 		if transactionLiked[txID] {
 			o = opinion.Like
 		}
-		consensusProvider.PayloadOpinionProvider.ProcessVote(&vote.OpinionEvent{
+		consensusProvider.ProcessVote(&vote.OpinionEvent{
 			ID:      transactionID,
 			Opinion: o,
 			Ctx:     vote.Context{Type: vote.ConflictType}})
 	}))
 
-	consensusProvider.PayloadOpinionProvider.VoteError().Attach(events.NewClosure(func(err error) {
+	consensusProvider.Events.Error.Attach(events.NewClosure(func(err error) {
 		t.Log("VoteError", err)
 	}))
 
@@ -203,7 +203,7 @@ func TestOpinionFormer(t *testing.T) {
 	LikedThreshold = 1 * time.Second
 	LocallyFinalizedThreshold = 2 * time.Second
 
-	consensusProvider := NewConsensusProvider()
+	consensusProvider := NewConsensusMechanism()
 
 	testTangle := tangle.New(tangle.Consensus(consensusProvider))
 	defer testTangle.Shutdown()
@@ -261,15 +261,15 @@ func TestOpinionFormer(t *testing.T) {
 		wg.Done()
 	}))
 
-	consensusProvider.PayloadOpinionProvider.Vote().Attach(events.NewClosure(func(transactionID string, initialOpinion opinion.Opinion) {
+	consensusProvider.Events.Vote.Attach(events.NewClosure(func(transactionID string, initialOpinion opinion.Opinion) {
 		t.Log("Voting requested for:", transactionID)
-		consensusProvider.PayloadOpinionProvider.ProcessVote(&vote.OpinionEvent{
+		consensusProvider.ProcessVote(&vote.OpinionEvent{
 			ID:      transactionID,
 			Opinion: opinion.Dislike,
 			Ctx:     vote.Context{Type: vote.ConflictType}})
 	}))
 
-	consensusProvider.PayloadOpinionProvider.VoteError().Attach(events.NewClosure(func(err error) {
+	consensusProvider.Events.Error.Attach(events.NewClosure(func(err error) {
 		t.Log("VoteError", err)
 	}))
 
@@ -282,6 +282,182 @@ func TestOpinionFormer(t *testing.T) {
 	wg.Wait()
 
 	t.Log("Waiting shutdown..")
+}
+
+func TestDeriveOpinion(t *testing.T) {
+	now := time.Now()
+
+	// A{t0, pending} -> B{t-t0 < c, Dislike, One}
+	{
+		conflictSet := ConflictSet{
+			OpinionEssence{
+				timestamp:        now,
+				liked:            false,
+				levelOfKnowledge: Pending,
+			}}
+
+		opinion := deriveOpinion(now.Add(1*time.Second), conflictSet)
+		assert.Equal(t, OpinionEssence{
+			timestamp:        now.Add(1 * time.Second),
+			liked:            false,
+			levelOfKnowledge: One,
+		}, opinion)
+	}
+
+	// A{t0, Like, One} -> B{c < t-t0 < c + d, Dislike, One}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            true,
+					levelOfKnowledge: One,
+				}}
+
+			opinion := deriveOpinion(now.Add(1*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(1 * time.Second),
+				liked:            false,
+				levelOfKnowledge: One,
+			}, opinion)
+		}
+	}
+
+	// {t0, Like, Two} -> {t-t0 > c + d, Dislike, Two}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            true,
+					levelOfKnowledge: Two,
+				}}
+
+			opinion := deriveOpinion(now.Add(1*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(1 * time.Second),
+				liked:            false,
+				levelOfKnowledge: Two,
+			}, opinion)
+		}
+	}
+
+	// A{t0, Dislike, Two}, B{t1, Dislike, Two} -> {t-t0 >> c + d, Dislike, Pending}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            false,
+					levelOfKnowledge: Two,
+				}, OpinionEssence{
+					timestamp:        now,
+					liked:            false,
+					levelOfKnowledge: Two,
+				}}
+
+			opinion := deriveOpinion(now.Add(1*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(1 * time.Second),
+				liked:            false,
+				levelOfKnowledge: Pending,
+			}, opinion)
+		}
+	}
+
+	// * double check this case
+	//  {t0, Dislike, One}, {t1, Dislike, One} -> {t-t0 > 0, Dislike, one}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            false,
+					levelOfKnowledge: One,
+				}, OpinionEssence{
+					timestamp:        now.Add(1 * time.Second),
+					liked:            false,
+					levelOfKnowledge: One,
+				}}
+
+			opinion := deriveOpinion(now.Add(10*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(10 * time.Second),
+				liked:            false,
+				levelOfKnowledge: One,
+			}, opinion)
+		}
+	}
+
+	// * double check this case
+	//  {t0, Like, One}, {t1, Dislike, One} -> {t - t0 > c, Dislike, one}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            true,
+					levelOfKnowledge: One,
+				}, OpinionEssence{
+					timestamp:        now.Add(1 * time.Second),
+					liked:            false,
+					levelOfKnowledge: One,
+				}}
+
+			opinion := deriveOpinion(now.Add(10*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(10 * time.Second),
+				liked:            false,
+				levelOfKnowledge: One,
+			}, opinion)
+		}
+	}
+
+	//  {t0, Dislike, Two}, {t1, Like, One} -> {t10, Dislike, one}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            false,
+					levelOfKnowledge: Two,
+				}, OpinionEssence{
+					timestamp:        now.Add(1 * time.Second),
+					liked:            true,
+					levelOfKnowledge: One,
+				}}
+
+			opinion := deriveOpinion(now.Add(6*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(6 * time.Second),
+				liked:            false,
+				levelOfKnowledge: One,
+			}, opinion)
+		}
+	}
+
+	//  {t0, Dislike, Two}, {t1, Dislike, One} -> {t10, Dislike, one}
+	{
+		{
+			conflictSet := ConflictSet{
+				OpinionEssence{
+					timestamp:        now,
+					liked:            false,
+					levelOfKnowledge: Two,
+				}, OpinionEssence{
+					timestamp:        now.Add(1 * time.Second),
+					liked:            false,
+					levelOfKnowledge: One,
+				}}
+
+			opinion := deriveOpinion(now.Add(6*time.Second), conflictSet)
+			assert.Equal(t, OpinionEssence{
+				timestamp:        now.Add(6 * time.Second),
+				liked:            false,
+				levelOfKnowledge: One,
+			}, opinion)
+		}
+	}
 }
 
 type wallet struct {
