+++ plugins/messagelayer/consensus.go
@@ -90,12 +90,12 @@ func configureConsensus(_ *node.Plugin) {
 	configureFPC()
 
 	// subscribe to FCOB events
-	Consensus().PayloadOpinionProvider.Vote().Attach(events.NewClosure(func(id string, initOpn opinion.Opinion) {
+	Consensus().Events.Vote.Attach(events.NewClosure(func(id string, initOpn opinion.Opinion) {
 		if err := Voter().Vote(id, vote.ConflictType, initOpn); err != nil {
 			log.Warnf("FPC vote: %s", err)
 		}
 	}))
-	Consensus().PayloadOpinionProvider.VoteError().Attach(events.NewClosure(func(err error) {
+	Consensus().Events.Error.Attach(events.NewClosure(func(err error) {
 		log.Errorf("FCOB error: %s", err)
 	}))
 
@@ -146,7 +146,7 @@ func configureFPC() {
 		log.Debugf("executed round with rand %0.4f for %d vote contexts on %d peers, took %v", roundStats.RandUsed, voteContextsCount, peersQueried, roundStats.Duration)
 	}))
 
-	Voter().Events().Finalized.Attach(events.NewClosure(Consensus().PayloadOpinionProvider.ProcessVote))
+	Voter().Events().Finalized.Attach(events.NewClosure(Consensus().ProcessVote))
 	Voter().Events().Finalized.Attach(events.NewClosure(func(ev *vote.OpinionEvent) {
 		if ev.Ctx.Type == vote.ConflictType {
 			log.Infof("FPC finalized for transaction with id '%s' - final opinion: '%s'", ev.ID, ev.Opinion)
